+++ xarray/tests/test_computation.py
@@ -852,6 +852,31 @@ def test_vectorize_dask_dtype():
     assert_identical(expected, actual)
     assert expected.dtype == actual.dtype
 
+    data_array = xr.DataArray([[0, 1, 2], [1, 2, 3]], dims=("x", "y"))
+    expected = data_array.copy()
+    actual = apply_ufunc(
+        identity,
+        data_array.chunk({"x": 1}),
+        vectorize=True,
+        dask="parallelized",
+    )
+
+    assert_identical(expected, actual)
+    assert expected.dtype == actual.dtype
+
+    data_array = xr.DataArray([[0, 1, 2], [1, 2, 3]], dims=("x", "y"))
+    expected = xr.DataArray([1, 2], dims=["x"])
+    actual = apply_ufunc(
+        pandas_median,
+        data_array.chunk({"x": 1}),
+        input_core_dims=[["y"]],
+        vectorize=True,
+        dask="parallelized",
+    )
+
+    assert_identical(expected, actual)
+    assert actual.dtype == 'float64'
+
 
 @pytest.mark.xfail(LooseVersion(dask.__version__) < "2.3", reason="dask GH5274")
 @requires_dask