+++ xarray/tests/test_sparse.py
@@ -884,6 +884,6 @@ def test_apply_ufunc_check_meta_coherence():
     sparse_meta = ac.data._meta
 
     # if dask computed meta, it would be np.ndarray
-    expected = xr.apply_ufunc(lambda x: x, ac, dask="parallelized").data._meta
+    result = xr.apply_ufunc(lambda x: x, ac, dask="parallelized").data._meta
 
-    assert_sparse_equal(expected, sparse_meta)
+    assert_sparse_equal(result, sparse_meta)